# backend/api/api.go
package api

import (
	"context"
	"strings"
	"time"
	"trxd/api/middlewares"
	"trxd/api/routes/attachments_create"
	"trxd/api/routes/attachments_delete"
	"trxd/api/routes/categories_create"
	"trxd/api/routes/categories_delete"
	"trxd/api/routes/categories_get"
	"trxd/api/routes/categories_update"
	"trxd/api/routes/challenges_all_get"
	"trxd/api/routes/challenges_create"
	"trxd/api/routes/challenges_delete"
	"trxd/api/routes/challenges_get"
	"trxd/api/routes/challenges_hidden"
	"trxd/api/routes/challenges_update"
	"trxd/api/routes/configs_get"
	"trxd/api/routes/configs_update"
	"trxd/api/routes/flags_create"
	"trxd/api/routes/flags_delete"
	"trxd/api/routes/flags_update"
	"trxd/api/routes/instances_create"
	"trxd/api/routes/instances_delete"
	"trxd/api/routes/instances_get"
	"trxd/api/routes/instances_update"
	"trxd/api/routes/stats"
	"trxd/api/routes/submissions_create"
	"trxd/api/routes/submissions_delete"
	"trxd/api/routes/submissions_get"
	"trxd/api/routes/teams_all_get"
	"trxd/api/routes/teams_get"
	"trxd/api/routes/teams_join"
	"trxd/api/routes/teams_join_get"
	"trxd/api/routes/teams_password"
	"trxd/api/routes/teams_register"
	"trxd/api/routes/teams_scoreboard"
	"trxd/api/routes/teams_scoreboard_graph"
	"trxd/api/routes/teams_search"
	"trxd/api/routes/teams_update"
	"trxd/api/routes/users_all_get"
	"trxd/api/routes/users_get"
	"trxd/api/routes/users_info"
	"trxd/api/routes/users_login"
	"trxd/api/routes/users_logout"
	"trxd/api/routes/users_password"
	"trxd/api/routes/users_register"
	"trxd/api/routes/users_role"
	"trxd/api/routes/users_search"
	"trxd/api/routes/users_update"
	"trxd/db"
	"trxd/utils"
	"trxd/utils/consts"
	"trxd/utils/log"

	_ "trxd/docs" // swag docs

	"github.com/gofiber/fiber/v2"
	"github.com/gofiber/fiber/v2/middleware/compress"
	"github.com/gofiber/fiber/v2/middleware/csrf"
	"github.com/gofiber/fiber/v2/middleware/favicon"
	"github.com/gofiber/fiber/v2/middleware/helmet"
	"github.com/gofiber/fiber/v2/middleware/monitor"
	"github.com/gofiber/swagger"
)

// @title TRXd API
// @BasePath /

var (
	noAuth = middlewares.NoAuth
	player = middlewares.Player
	author = middlewares.Author
	admin  = middlewares.Admin

	team = middlewares.Team

	start = middlewares.Start
	end   = middlewares.End
)

func SetupApp(ctx context.Context) *fiber.App {
	app := fiber.New(fiber.Config{
		AppName:   consts.Name,
		BodyLimit: 200 * 1024 * 1024, // 200MB
	})

	SetupFeatures(app)
	SetupApi(ctx, app)

	app.Static("/", "./frontend")

	app.Use("/attachments", player, team, start, middlewares.Attachments)
	app.Static("/attachments", "./attachments", fiber.Static{
		Download: true,
	})

	// serve index.html for all non-API routes so SvelteKit's
	// client-side router handles them. Only /api/* returns a JSON 404.
	app.Use(func(c *fiber.Ctx) error {
		if strings.HasPrefix(c.Path(), "/api/") {
			return utils.Error(c, fiber.StatusNotFound, consts.NotFound)
		}
		return c.SendFile("./frontend/index.html")
	})

	return app
}

func Shutdown(app *fiber.App) {
	err := app.Shutdown()
	if err != nil {
		log.Error("Failed to shutdown Fiber app:", "err", err)
	}
}

func SetupFeatures(app *fiber.App) {
	if consts.AntiPanic {
		app.Use(func(c *fiber.Ctx) error {
			defer func() {
				r := recover()
				if r == nil {
					return
				}
				log.Critical("Panic recovered:", "crit", r)
				_ = utils.Error(c, fiber.StatusInternalServerError, consts.InternalServerError)
			}()
			return c.Next()
		})
	}

	app.Use(compress.New())

	app.Use(csrf.New(csrf.Config{
		KeyLookup:         "header:X-CSRF-Token",
		CookieSameSite:    fiber.CookieSameSiteLaxMode,
		CookieSessionOnly: true,
		Expiration:        48 * time.Hour,
		Session:           db.Store,
	}))

	app.Use(helmet.New(helmet.Config{
		ContentSecurityPolicy:     "",
		CrossOriginEmbedderPolicy: "unsafe-none",
	}))

	app.Use(favicon.New(favicon.Config{
		File: "./static/favicon.ico",
		URL:  "/favicon.ico",
	}))

	app.Get("/monitor", admin, monitor.New(monitor.Config{
		Title: consts.Name + " Monitor",
	}))

	// app.Static("/swagger/doc.json", "./docs/swagger.json")
	// app.Get("/swagger/*", admin, swagger.HandlerDefault)
	app.Get("/swagger/*", swagger.HandlerDefault)
}

func SetupApi(ctx context.Context, app *fiber.App) {
	mode, err := db.GetConfig(ctx, "user-mode")
	if err != nil {
		log.Error("Failed to get user-mode config:", "err", err)
		mode = "false"
	}
	userMode := mode == "true"

	var api fiber.Router
	if log.GetLevel() == log.DebugLevel {
		api = app.Group("/api", middlewares.Debug)
	} else {
		api = app.Group("/api")
	}

	api.Post("/register", noAuth, users_register.Route)
	api.Post("/login", noAuth, users_login.Route)
	api.Post("/logout", noAuth, users_logout.Route)
	api.Get("/info", noAuth, users_info.Route)
	api.Get("/scoreboard", noAuth, teams_scoreboard.Route)
	api.Get("/scoreboard/graph", noAuth, teams_scoreboard_graph.Route)

	api.Patch("/users", player, users_update.Route)
	api.Patch("/users/role", admin, users_role.Route)
	api.Patch("/users/password", player, users_password.Route)
	if !userMode {
		api.Get("/users", noAuth, users_all_get.Route)
		api.Get("/users/search", author, users_search.Route)
		api.Get("/users/:id", noAuth, users_get.Route)
	}

	if !userMode {
		api.Post("/teams/register", player, teams_register.Route)
		api.Post("/teams/join", player, teams_join.Route)
		api.Get("/teams/join", player, teams_join_get.Route)
		api.Patch("/teams", player, team, teams_update.Route)
		api.Patch("/teams/password", player, team, teams_password.Route)
	}
	api.Get("/teams", noAuth, teams_all_get.Route)
	api.Get("/teams/search", author, teams_search.Route)
	api.Get("/teams/:id", noAuth, teams_get.Route)

	api.Post("/categories", author, categories_create.Route)
	api.Patch("/categories", author, categories_update.Route)
	api.Delete("/categories", author, categories_delete.Route)
	api.Get("/categories", player, team, start, categories_get.Route)

	api.Post("/challenges", author, challenges_create.Route)
	api.Patch("/challenges", author, challenges_update.Route)
	api.Patch("/challenges/hidden", author, challenges_hidden.Route)
	api.Delete("/challenges", author, challenges_delete.Route)
	api.Get("/challenges", player, team, start, challenges_all_get.Route)
	api.Get("/challenges/:id", player, team, start, challenges_get.Route)

	api.Post("/instances", player, team, start, instances_create.Route)
	api.Patch("/instances", player, team, start, instances_update.Route)
	api.Delete("/instances", player, team, start, instances_delete.Route)
	api.Get("/instances", author, instances_get.Route)

	api.Post("/submissions", player, team, start, end, submissions_create.Route)
	api.Get("/submissions", author, submissions_get.Route)
	api.Delete("/submissions", admin, submissions_delete.Route)

	api.Post("/attachments", author, attachments_create.Route)
	api.Delete("/attachments", author, attachments_delete.Route)

	api.Post("/flags", author, flags_create.Route)
	api.Patch("/flags", author, flags_update.Route)
	api.Delete("/flags", author, flags_delete.Route)

	api.Get("/configs", admin, configs_get.Route)
	api.Patch("/configs", admin, configs_update.Route)

	api.Get("/stats", author, stats.Route)
}