# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "TRXd API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/api/attachments": {
            "post": {
                "description": "Requires **Author** privileges or higher.\nCreates new attachments for an existing challenge with the provided files.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "[Author+] Creates new attachments for a challenge",
                "parameters": [
                    {
                        "description": "all fields are required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attachments_create.Data"
                        }
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "the list of files to upload (at least one file is required)",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid multipart form` | `Invalid form data` | `Missing required fields` | `ChallID must be at least 0` | `Attachments[i] must not exceed 128`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Possible errors: `Challenge not found`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "409": {
                        "description": "Possible errors: `Attachment already exists`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error fetching challenge` | `Error creating attachments`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Requires **Author** privileges or higher.\nDeletes attachments for an existing challenge with the provided names.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "[Author+] Deletes attachments for a challenge",
                "parameters": [
                    {
                        "description": "all fields are required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attachments_delete.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid JSON format` | `Missing required fields` | `ChallID must be at least 0` | `Names[i] must not exceed 128`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Possible errors: `Attachment not found`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error fetching attachment` | `Error deleting attachment`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "description": "Requires **Player** privileges or higher (with role **Player** is also required to be in a team and the competition has to be active).\nRetrieves a list of all categories (with role **Player** only visible one are returned).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "[Player+] Gets all categories",
                "responses": {
                    "200": {
                        "description": "List of category names",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error fetching categories`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Requires **Author** privileges or higher.\nCreates a new category with the provided name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "[Author+] Creates a new category",
                "parameters": [
                    {
                        "description": "all fields are required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/categories_create.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid JSON format` | `Missing required fields` | `Name must not exceed 32`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "409": {
                        "description": "Possible errors: `Category already exists`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error creating category`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Requires **Author** privileges or higher.\nDeletes an existing category with the provided name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "[Author+] Deletes a category",
                "parameters": [
                    {
                        "description": "all fields are required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/categories_delete.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid JSON format` | `Missing required fields` | `Name must not exceed 32`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error deleting category`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "patch": {
                "description": "Requires **Author** privileges or higher.\nUpdates an existing category with the provided name (also updates its own challenges categories).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "[Author+] Updates a category",
                "parameters": [
                    {
                        "description": "all fields are required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/categories_update.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid JSON format` | `Missing required fields` | `Name must not exceed 32` | `NewName must not exceed 32`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Possible errors: `Category not found`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "409": {
                        "description": "Possible errors: `Category already exists`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error fetching category` | `Error creating category` | `Error updating category` | `Error deleting category`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/challenges": {
            "get": {
                "description": "Requires **Player** privileges or higher (with role **Player** is also required to be in a team and the competition has to be active).\nRetrieves a list of all challenges (with role **Player** only visible one are returned).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "[Player+] Gets all challenges",
                "responses": {
                    "200": {
                        "description": "List of challenges with all dysplayable details (except solves)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/challenges_all_get.Chall"
                            }
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error fetching challenges`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Requires **Author** privileges or higher.\nCreates a new challenge with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "[Author+] Creates a new challenge",
                "parameters": [
                    {
                        "description": "all fields are required except **description**",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/challenges_create.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid JSON format` | `Missing required fields` | `Name must not exceed 32` | `Category must not exceed 32` | `Description must not exceed 10240` | `Type must be one of: Normal Container Compose` | `MaxPoints must be at least 0` | `ScoreType must be one of: Static Dynamic`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Possible errors: `Category not found`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "409": {
                        "description": "Possible errors: `Challenge already exists`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error creating challenge`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Requires **Author** privileges or higher.\nDeletes a challenge with the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "[Author+] Deletes a challenge",
                "parameters": [
                    {
                        "description": "all fields are required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/challenges_delete.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid JSON format` | `Missing required fields` | `ChallID must be at least 0`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error deleting challenge`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "patch": {
                "description": "Requires **Author** privileges or higher.\nUpdates an existing challenge with the provided fields, only the `chall_id` is required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "[Author+] Updates details of an existing challenge",
                "parameters": [
                    {
                        "description": "`chall_id` is required, the rest of the fields are optional, only provide the fields you want to update",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/challenges_update.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid JSON format` | `Missing required fields` | `No data provided to update` | `ChallID must be at least 0` | `Name must not exceed 128` | `Category must not exceed 32` | `Description must not exceed 10240` | `Authors[i] must not exceed 64` | `Tags[i] must not exceed 32` | `Type must be one of: Normal Container Compose` | `MaxPoints must be at least 0` | `ScoreType must be one of: Static Dynamic` | `Port must be at least 0` | `Port must not exceed 65535` | `ConnType must be one of: NONE TCP HTTP HTTPS` | `Lifetime must be at least 0` | `Invalid environment variables` | `MaxMemory must be at least 0` | `Invalid Max CPU, must be a positive 32-bit integer`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Possible errors: `Challenge not found` | `Category not found`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "409": {
                        "description": "Possible errors: `Challenge name already exists`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error fetching challenge` | `Error updating challenge`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/challenges/:id": {
            "get": {
                "description": "Requires **Player** privileges or higher (with role **Player** is also required to be in a team and the competition has to be active).\nRetrieves challenge details (with role **Player** only solves if the challenge is visible).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "[Player+] Gets challenge details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Challenge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "solves is only included if the requester has Player role, otherwise all the challenge details are included except the solves",
                        "schema": {
                            "$ref": "#/definitions/challenges_get.Chall"
                        }
                    },
                    "400": {
                        "description": "Possible errors: `Invalid challenge ID, must be non negative` | `id must be at least 0`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Possible errors: `Challenge not found`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error fetching challenge`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/challenges/hidden": {
            "patch": {
                "description": "Requires **Author** privileges or higher.\nToggles the hidden status of the specified challenges.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "[Author+] Toggles a list of challenges between hidden and visible",
                "parameters": [
                    {
                        "description": "a list of challenge IDs to toggle hidden status",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/challenges_hidden.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Possible errors: `Invalid JSON format` | `Missing required fields` | `ChallIDs[i] must be at least 0`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error updating challenge`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
        "/api/stats": {
            "get": {
                "description": "Requires **Author** privileges or higher.\nRetrieves a list useful stats.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "[Author+] Gets stats",
                "responses": {
                    "200": {
                        "description": "List of stats",
                        "schema": {
                            "$ref": "#/definitions/sqlc.GetAdminStatsRow"
                        }
                    },
                    "500": {
                        "description": "Possible errors: `Error fetching stats`",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "attachments_create.Data": {
            "type": "object",
            "required": [
                "chall_id"
            ],
            "properties": {
                "chall_id": {
                    "type": "integer"
                }
            }
        },
        "attachments_delete.Data": {
            "type": "object",
            "required": [
                "chall_id",
                "names"
            ],
            "properties": {
                "chall_id": {
                    "type": "integer"
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "categories_create.Data": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "categories_delete.Data": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "categories_update.Data": {
            "type": "object",
            "required": [
                "name",
                "new_name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "new_name": {
                    "type": "string"
                }
            }
        },
...
        "challenges_create.Data": {
            "type": "object",
            "required": [
                "category",
                "max_points",
                "name",
                "score_type",
                "type"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "max_points": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "score_type": {
                    "$ref": "#/definitions/sqlc.ScoreType"
                },
                "type": {
                    "$ref": "#/definitions/sqlc.DeployType"
                }
            }
        },
        "challenges_delete.Data": {
            "type": "object",
            "required": [
                "chall_id"
            ],
            "properties": {
                "chall_id": {
                    "type": "integer"
                }
            }
        },
...
        "challenges_hidden.Data": {
            "type": "object",
            "required": [
                "chall_ids"
            ],
            "properties": {
                "chall_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "models.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "sqlc.ScoreType": {
            "type": "string",
            "enum": [
                "Static",
                "Dynamic"
            ],
            "x-enum-varnames": [
                "ScoreTypeStatic",
                "ScoreTypeDynamic"
            ]
        },
...
    }
}